oesp_sdk_python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "oesp-sdk"
version = "0.2.0"
description = "Offline Exchange Secure Protocol (OESP) SDK for Python"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
authors = [
    { name = "OESP Team", email = "dev@oesp.protocol" }
]
keywords = ["oesp", "crypto", "offline", "exchange", "protocol", "security"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "PyNaCl>=1.5.0",
    "cryptography>=43.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["oesp_sdk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
